% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy}
\alias{toy}
\title{Toy dataset from Momocs}
\format{From Momocs:
\itemize{
\item bot
\item chaff
\item charring
\item flower
\item hearts
\item molars
\item mosquito
\item oak
\item olea
\item shapes
\item trilo
\item wings
}}
\source{
\link{Momocs}
}
\description{
Toy dataset from Momocs
}
