% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momecs-package.R
\docType{package}
\name{Momecs-package}
\alias{Momecs-package}
\title{Momocs}
\description{
The goal of Momecs is to ease exploration and hypotheses testing
on multivariate datasets, notably for morphometrics data.
It includes the most common approaches, namely:
PCA (Principal Component analysis), LDA (Linear Discriminant Analysis),
and friends. It essentially wraps on existing functions.
}
\details{
To cite Momecs in publications: not yet published.
}
\references{
Not yet published.
}
\seealso{
Have a look to:
\itemize{
\item Homepage: (http://momx.github.io/Momecs/)
\item Report issues and/or ask for features: (https://github.com/MomX/Momecs/issues)
\item Other MomX packages (https://MomX.github.io/MomX)
\item Contact: \code{bonhomme.vincent@gmail.com}
}
}
