% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{Momecs}
\alias{Momecs}
\title{Morphometric (and multivariate) exploration controlled by shiny}
\usage{
Momecs(x)
}
\arguments{
\item{x}{a Coe or PCA object. If missing, \link{toy} datasets are used}
}
\description{
Morphometric (and multivariate) exploration controlled by shiny
}
\examples{
\dontrun{
# on toy dataset(s)
Momecs()

# passing a Coe
h <- hearts \%>\% efourier(3)
Momecs(h)
}
}
