% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{Momecs}
\alias{Momecs}
\title{Morphometric (or multivariate) exploration controlled by shiny}
\usage{
Momecs(x = toy, launch.brower = FALSE)
}
\arguments{
\item{x}{a Coe or PCA object}

\item{launch.brower}{\code{logical} to feed \link[shiny:runApp]{shiny::runApp}, whether to
launch the app within your default browser (default: \code{FALSE})}
}
\description{
Morphometric (or multivariate) exploration controlled by shiny
}
\examples{
\dontrun{
Momecs()
hearts \%>\% efourier(3)  \%>\% Momecs()
}
}
